Added command line options to control compatibility hacks.
Enabled overrides for the date and the time, which is useful for
bypassing "prime time" hours when the cave is closed, proving you're
a wizard, avoiding Y2K bugs, and getting a reproducible seed for the
random number generator.

Added an option to disable the Y2K hack (which is enabled by default).

Extended the Wizard spreadsheet to show the magic word calculation if
you override the clock to 1977-01-01 00:00.

Added options to get a summary of the options.