Reworked how DATA statement variable lists work.
Motivated by CROW0000, which uses index controls in several DATA
statements.

DATA statement variable lists are restricted versions of an IO item
list.  I tried to reconcile that again, but the fact that the index
controls in DATA statements have to be evaluated at translation time
and those in IO lists have to be evaluated at run time made that
difficult.  So in the end I extended variable lists.

Verified that WOOD0350v2 works as before.

The need for this was discovered after starting to see if I can get
the REAL type working (because CROW0000 uses a real-valued random
number generator from the system).  The REAL work is just starting.

I also discovered some missed opportunities to check syntax when
parsing FORMAT statements.